提供数据源获取和注册服务

+++ src/Dist.Product.Dme/Dist.Dme.Model/DTO/DatasourceAddDTO.cs
@@ -13,7 +13,7 @@ namespace Dist.Dme.Model.DTO
         // public int IsLocal { get; set; }
         public String Type { get; set; }
         public String Connection { get; set; }
-        public int CreateTime { get; set; }
+        public long CreateTime { get; set; }
         public String Remark { get; set; }
     }
 }
